+++ .env
@@ -1,1 +1,2 @@
 NODE_PATH=src/
+SKIP_PREFLIGHT_CHECK=true
