+++ .eslintrc.js
@@ -1,0 +1,13 @@
+module.exports = {
+  parser: 'babel-eslint',
+  plugins: ['react'],
+  extends: [
+    '@sheerun/eslint-config-standard/index.js',
+    'plugin:react/recommended',
+  ],
+  rules: {
+    'react/prop-types': 'off',
+    'react/display-name': 'off',
+    'comma-dangle': 'off',
+  },
+}
