+++ src/ducks/Studio/Tabs/UniswapHistoricalBalance/UniswapHistoricalBalance.module.scss
@@ -7,7 +7,7 @@
 .header {
   display: flex;
   justify-content: space-between;
-  padding: 8px 8px 0;
+  padding: 8px 8px 5px;
 }
 
 .chart {
